web/server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.1.2
python-multipart==0.0.9
httpx>=0.26.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
PyJWT>=2.8.0
aiohttp==3.9.3
discord.py==2.3.2
alembic==1.13.1
